+++ modules/www/fluent9-tools/src/ConfirmationDialog.tsx
@@ -12,71 +12,7 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@fluentui/react-components';
-import {isString} from '@freik/typechk';
-
-export type TextInputProps = {
-  data: DialogData;
-  onConfirm: (value: string) => void;
-  title: string;
-  text: string;
-  initialValue: string;
-  yesText?: string;
-  noText?: string;
-  minWidth?: number;
-  maxWidth?: number;
-};
-
-export function TextInput({
-  data: [hidden, hide],
-  onConfirm,
-  title,
-  text,
-  initialValue,
-  yesText,
-  noText,
-  minWidth,
-  maxWidth,
-}: TextInputProps): React.JSX.Element {
-  const [input, setInput] = useState(initialValue);
-  const confirmAndClose = () => {
-    hide();
-    onConfirm(input);
-  };
-  const yes = yesText ?? 'Yes';
-  const no = noText ?? 'No';
-  const dlgContentProps = {
-    type: DialogType.normal,
-    title,
-    closeButtonAriaLabel: 'Close',
-    subText: text,
-  };
-  return (
-    <Dialog
-      hidden={hidden}
-      onDismiss={hide}
-      minWidth={minWidth}
-      maxWidth={maxWidth}
-      dialogContentProps={dlgContentProps}
-    >
-      <Stack>
-        <Text>{text}</Text>
-        <TextField
-          value={input}
-          onChange={(ev, newValue) => setInput(newValue ?? initialValue)}
-        />
-        <br />
-      </Stack>
-      <DialogFooter>
-        <PrimaryButton style={{ float: 'left' }} onClick={hide}>
-          {no}
-        </PrimaryButton>
-        <DefaultButton style={{ float: 'right' }} onClick={confirmAndClose}>
-          {yes}
-        </DefaultButton>
-      </DialogFooter>
-    </Dialog>
-  );
-}
+import { isString } from '@freik/typechk';
 
 export type ConfirmationDialogProps = {
   data: DialogData;
@@ -100,11 +36,9 @@ export function ConfirmationDialog({
 }: ConfirmationDialogProps): React.JSX.Element {
   const yesEl = yes ?? 'Yes';
   const noEl = no ?? 'No';
-  const openEl = isString(open) ? (<Button>{open}</Button>) : open;
+  const openEl = isString(open) ? <Button>{open}</Button> : open;
   return (
-    <Dialog
-      {...props}
-    >
+    <Dialog {...props}>
       <DialogTrigger>{openEl}</DialogTrigger>
       <DialogSurface>
         <DialogBody>
