+++ modules/www/fluent9-tools/src/TextInputDialog.tsx
@@ -1,0 +1,63 @@
+export type TextInputProps = {
+  data: DialogData;
+  onConfirm: (value: string) => void;
+  title: string;
+  text: string;
+  initialValue: string;
+  yesText?: string;
+  noText?: string;
+  minWidth?: number;
+  maxWidth?: number;
+};
+
+export function TextInput({
+  data: [hidden, hide],
+  onConfirm,
+  title,
+  text,
+  initialValue,
+  yesText,
+  noText,
+  minWidth,
+  maxWidth,
+}: TextInputProps): React.JSX.Element {
+  const [input, setInput] = useState(initialValue);
+  const confirmAndClose = () => {
+    hide();
+    onConfirm(input);
+  };
+  const yes = yesText ?? 'Yes';
+  const no = noText ?? 'No';
+  const dlgContentProps = {
+    type: DialogType.normal,
+    title,
+    closeButtonAriaLabel: 'Close',
+    subText: text,
+  };
+  return (
+    <Dialog
+      hidden={hidden}
+      onDismiss={hide}
+      minWidth={minWidth}
+      maxWidth={maxWidth}
+      dialogContentProps={dlgContentProps}
+    >
+      <Stack>
+        <Text>{text}</Text>
+        <TextField
+          value={input}
+          onChange={(ev, newValue) => setInput(newValue ?? initialValue)}
+        />
+        <br />
+      </Stack>
+      <DialogFooter>
+        <PrimaryButton style={{ float: 'left' }} onClick={hide}>
+          {no}
+        </PrimaryButton>
+        <DefaultButton style={{ float: 'right' }} onClick={confirmAndClose}>
+          {yes}
+        </DefaultButton>
+      </DialogFooter>
+    </Dialog>
+  );
+}
