+++ package.json
@@ -28,7 +28,7 @@
         "@eslint/eslintrc": "^3.3.5",
         "@types/bun": "^1.3.14",
         "@types/debug": "^4.1.13",
-        "@types/node": "^25.9.2",
+        "@types/node": "^25.9.3",
         "@types/react-test-renderer": "^19.1.0",
         "@typescript-eslint/eslint-plugin": "^8.61.0",
         "@typescript-eslint/parser": "^8.61.0",
@@ -55,7 +55,7 @@
         "react-dom": "^19.2.7",
         "react-scripts": "^5.0.1",
         "@freik/workspace": "^0.8.1",
-        "esbuild": "^0.28.0",
+        "esbuild": "^0.28.1",
         "madge": "^8.0.0",
         "prettier": "^3.8.4"
       }
@@ -87,7 +87,7 @@
     "@eslint/eslintrc": "^3.3.5",
     "@types/bun": "^1.3.14",
     "@types/debug": "^4.1.13",
-    "@types/node": "^25.9.2",
+    "@types/node": "^25.9.3",
     "@types/react-test-renderer": "^19.1.0",
     "@typescript-eslint/eslint-plugin": "^8.61.0",
     "@typescript-eslint/parser": "^8.61.0",
@@ -114,7 +114,7 @@
     "react-dom": "^19.2.7",
     "react-scripts": "^5.0.1",
     "@freik/workspace": "^0.8.1",
-    "esbuild": "^0.28.0",
+    "esbuild": "^0.28.1",
     "madge": "^8.0.0",
     "prettier": "^3.8.4"
   },
